+++ packages/vite-plugin-vui/src/plugin.ts
@@ -151,6 +151,7 @@ export async function viteVuiPlugin(
       hinttip: 'mdi-help-circle-outline' as any,
       clear: 'mdi-close' as any,
       reload: 'mdi-reload' as any,
+      fileUpload: 'mdi-file-upload' as any,
       // navigationExpand: (gen, active) => {
       //   return gen({
       //     name: 'mdi-menu-down' as any,